+++ Backend/Infrastructure/Interfaces/IRepository.cs
@@ -7,7 +7,7 @@ public interface IRepository<T>
     List<T> All();
     T Create(T t);
     bool Delete(int id);
-    T Single(long id);
+    T Single(int id);
     T Update(T model);
 }
